% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccumulatedWhales.R
\name{accum_fun}
\alias{accum_fun}
\title{Function to calculate the accumulated number of whales using previously estimated regression coefficients}
\usage{
accum_fun(
  a = -77.41,
  b = 0.03234,
  c = 0.1605,
  d = -0.0003323,
  mu = 60,
  sigma = sqrt(60),
  x = 320,
  ReportList = FALSE
)
}
\arguments{
\item{a}{Parameter from log-link GLM representing the intercept}

\item{b}{Parameter from log-link GLM representing the categorical effect of year}

\item{c}{Parameter from log-link GLM representing the effect of day}

\item{d}{Parameter from log-link GLM representing the effect of day^2}

\item{mu}{Mean of normal distribution representing time whale remains in area}

\item{sigma}{Standard deviation of normal distribution representing time whale remains in area}

\item{x}{Day of year to calculated accumulated number of whales: Assuming 320 because after 320 no new whales come in}

\item{year}{Year for calculating accumulated number of whales}
}
\description{
Function to calculate the accumulated number of whales using previously estimated regression coefficients
}
\examples{
accum_fun(x = 320)
}
