% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SIR.R
\name{CALC.ANALYTIC.Q.MVLNORM}
\alias{CALC.ANALYTIC.Q.MVLNORM}
\title{Compute analytic estimates of q, the scaling parameter between indices and
absolute population size}
\usage{
CALC.ANALYTIC.Q.MVLNORM(
  rel.abundance,
  rel.var.covar,
  rel.hess,
  Pred_N,
  start_yr,
  add_CV = 0,
  beta,
  num.IA
)
}
\arguments{
\item{rel.abundance}{Relative abundance index}

\item{rel.var.covar}{Variance covariance (narrow)}

\item{rel.hess}{Hessian matrix of relative abundance (narrow)}

\item{Pred_N}{Predicted population}

\item{start_yr}{Initial year}

\item{add_CV}{Coefficient of variation}

\item{beta}{Density dependent catchability coefficient I = q*N^(1+beta)}

\item{num.IA}{Index of abundance}
}
\value{
A numeric estimator for $q$.
}
\description{
Compute analytic estimates of q, the scaling parameter between indices and
absolute population size
}
