% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SIR.R
\name{LNLIKE.MVLNORM.IAs}
\alias{LNLIKE.MVLNORM.IAs}
\title{Compute the log-likelihood of indices of abundance assuming multivariate lognormal}
\usage{
LNLIKE.MVLNORM.IAs(
  rel.abundance,
  rel.var.covar,
  Pred_N,
  start_yr,
  q.values,
  add.CV,
  log = TRUE
)
}
\arguments{
\item{rel.abundance}{Relative abundance indices}

\item{rel.var.covar}{Variance covariance matrix}

\item{Pred_N}{Predicted population size}

\item{start_yr}{Initial year}

\item{q.values}{Scaling parameter}

\item{add.CV}{Coefficient of variation [UNUSED]}

\item{log}{Boolean, return log likelihood (default TRUE) or
likelihood.}
}
\value{
List of likelihood based on multivariate lognormal distribution
}
\description{
Compute the log-likelihood of indices of abundance assuming multivariate lognormal
}
