% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmsy_to_z_functions.R
\name{pmsy_z_haider}
\alias{pmsy_z_haider}
\title{Function to convert between Pmsy and Z for Haider et al 2017 depensation function}
\usage{
pmsy_z_haider(z, Pmsy, k = 100, r = 0.1, q = 0.02)
}
\arguments{
\item{q}{}
}
\value{

}
\description{

}
