% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmsy_to_z_functions.R
\name{pmsy_z_linli}
\alias{pmsy_z_linli}
\title{Function to convert between Pmsy and Z for Lin and Li 2002 depensation function}
\usage{
pmsy_z_linli(z, Pmsy, k = 100, r = 0.1, q = 0.02)
}
\arguments{
\item{q}{}
}
\value{

}
\description{
report{Lin2002,
  abstract = {In this study an attempt is made to investigate comprehensively the maximum sustainable yield (MSY) of Allee population dynamic system. The results show that: (1) because the net production rate of Allee system is a function of the shape parameter d, all Allee systems which possess different B (B=A/K, here K is the carrying capacity, and A the parameter of Allee effect) are adapted themselves to the environments by controlling their production rate. Allee effect does not change the interaction between the system and environment, but it will reduce the MSY; (2) the MSY of Allee system is approximately one-nineteenth of the height of the allometry curve of body size for all A/K50.01. However when the net production rate R takes the values from 3 to 15, the MSY of Allee system is also approximately 1/19 of the height of the allometry curve of body size for all A/K 50.001.},
  author = {Zhen-Shan Lin and Bai-Lian Li},
  journal = {Ecological Modelling},
  keywords = {Allee-effect,MSY wwwelseviercom/locate/ecolmodel,Population dynamics},
  pages = {1-7},
  title = {The maximum sustainable yield of Allee dynamic system},
  volume = {154},
  url = {www.elsevier.com/locate/ecolmodel},
  year = {2002},
}
}
