% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_specification.R
\name{sir_control}
\alias{sir_control}
\title{Set parameters for SIR}
\usage{
sir_control(
  K_bisect_lims = c(1, 5e+05),
  K_bisect_tol = 1e-04,
  threshold = 1e-17,
  progress_bar = FALSE,
  verbose = 0
)
}
\arguments{
\item{K_bisect_lims}{Bounds for K when performing the bisection method of
Pund and Butterworth (1995). Two-element vector denoting upper and lower
bounds. Default is \code{c(1, 5e5)}.}

\item{K_bisect_tol}{Tolerance passed to \code{uniroot} when using the
bisection method to find K.}

\item{threshold}{Threshold for McCallister et al. (1994) SIR. Data specific.
Defaults to 1e-17.}

\item{progress_bar}{Logical; display progress bar?}

\item{verbose}{Integer, defaults to 0; See Details for specifics. Note that
  more intermediate output will slow down sampling.

Use of \code{progress_bar} and \code{verbose} together is not allowed. If
both are set, only \code{progress_bar} will be shown. The \code{verbose}
argument takes an integer, controlling how much is output. All output uses
\code{message}, so works with functions like \code{supressMessages}. These
levels are

\itemize{
  \item \code{0} no output;
  \item \code{1} sample count and MVP violations;
  \item \code{2} above plus likelihood totals;
  \item \code{3} above plus likelihood components;
  \item \code{4} above plus sampled parameter values.
intermediate

}}
}
\value{
A list to be passed to the \code{control} argument of \code{StateSpaceSIR}.
}
\description{
Set fitting parameters for SIR with reasonable defaults.
}
