% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SIR.R
\name{StateSpaceSIR}
\alias{StateSpaceSIR}
\title{StateSpaceSIR controls the sampling from the priors, the bisection and
likelihoods and the output functions}
\usage{
StateSpaceSIR(
  file_name = "NULL",
  allee_model = 0,
  n_resamples = 1000,
  priors = make_prior_list(),
  catch_multipliers = make_multiplier_list(),
  target.Yr = 2008,
  num.haplotypes = 66,
  output.Yrs = c(2008),
  abs.abundance = Abs.Abundance,
  abs.abundance.key = TRUE,
  rel.abundance = Rel.Abundance,
  rel.abundance.key = TRUE,
  count.data = NULL,
  count.data.key = FALSE,
  min_rmax = 0.118,
  growth.rate.obs = c(0.074, 0.033, TRUE),
  growth.rate.Yrs = c(1995, 1996, 1997, 1998),
  catch.data = Catch.data,
  realized_prior = FALSE,
  control = sir_control()
)
}
\arguments{
\item{file_name}{name of a file to identified the files exported by the
function#'}

\item{allee_model}{Switch to determine the depensation model. 0 = no Allee effect; 1 = Hilborn et al 2014 P50 Allee Effect; 2 = Logistic Allee effect; 3 = Lin and Li 2002; 4 = Haider et al 2017. The depensation parameter is defined by \code{P50} using \link{make_prior_list}.}

\item{n_resamples}{number of resamples to compute the marginal posterior
distributions}

\item{priors}{List of priors, usually generated using \link{make_prior_list}.
Default is the default of \code{make_prior_list}. See details.}

\item{catch_multipliers}{List of catch multipliers, generated using \link{make_multiplier_list}
Can either be estimated or explicitly provided. Default is \code{make_multiplier_list}.}

\item{target.Yr}{year of the target population estimate for the bisection
method. Default is 2008}

\item{num.haplotypes}{number of haplotypes to compute minimum viable
population (from Jackson et al., 2006 and IWC, 2007)}

\item{output.Yrs}{year for outputing the predicted abundance estimates.
Default is 2008, but multiple years can be specified. For example, if
outputs for 2005 and 2008 are needed, output.Yrs = c(2005, 2008)}

\item{abs.abundance}{R object containing year, estimate of absolute
abundance, and CV (see example)}

\item{abs.abundance.key}{key to speficy if absolute abundance data are used
in the likelihood. Default is TRUE}

\item{rel.abundance}{data.frame of relative abundance indices. Columns are Index, Year, IA.obs, and Variance-Covariance Matrix of Logs (column names do not matter for var-covar)}

\item{rel.abundance.key}{key to speficy if relative abundance data are used
in the likelihood. Default is TRUE}

\item{count.data}{R object containing years, estimates of counts and effort.
NOT USED}

\item{count.data.key}{key to speficy in count data are used. Default is
FALSE. NOT USED}

\item{min_rmax}{Upper bound on rmax sampling (default to 0.118)}

\item{growth.rate.obs}{observed growth rate (1st element) and standard error
(2nd element) as in Zerbni et al. (2011). If third element is FALSE, the
growth rate is not included in the likelihood}

\item{growth.rate.Yrs}{Years for which the growth.rate.obs were computed (as
in Zerbini et al., 2011)}

\item{catch.data}{R list where each layer contains a data.frame of the Year, Min Catch, and Max Catch of that catch time-series.}

\item{realized_prior}{Key to specify if realized prior is to be extracted. Default is FALSE.}

\item{control}{A list of control parameters, usually generated by
\code{sir_control}.}
}
\value{
A \code{list} containing posterior samples and metadata

Current default prior specification:
\code{
make_prior_list(r_max = make_prior(runif, 0, 0.118),
                K = make_prior(use = FALSE),
                var_N = make_prior(0),
                N_obs = make_prior(runif, 500, 40000),
                add_CV = make_prior(use = FALSE),
                add_VAR_IA = make_prior(use = FALSE),
                catch_sample = make_prior(runif, 0, 1),
                z = make_prior(2.39),
                Pmsy = make_prior(use = FALSE),
                P50 = make_prior(0),
                q_IA1 = make_prior(use = FALSE),
                q_IA2 = make_prior(0),
                q_count = make_prior(use = FALSE)

 make_multiplier_list(c_mult_1 = make_prior(1))}
}
\description{
StateSpaceSIR controls the sampling from the priors, the bisection and
likelihoods and the output functions
}
\examples{

\dontrun{
StateSpaceSIR(file_name = "test.N2005",
             allee_model = 0,
             n_resamples = 100,
             priors = make_prior_list(),
             catch_multipliers = make_multiplier_list(),
             Klim = c(1, 500000),
             target.Yr = 2005,
             num.haplotypes = 0,
             tolerance.for.bisection = 0.0001,
             output.Yrs = c(2005, 2006),
             abs.abundance = Abs.Abundance.2005,
             abs.abundance.key = TRUE,
             rel.abundance = Rel.Abundance,
             rel.abundance.key = TRUE,
             count.data = Count.Data,
             count.data.key = FALSE,
             growth.rate.obs = c(0.074, 0.033, TRUE),
             growth.rate.Yrs = c(1995, 1996, 1997, 1998),
             catch.data = Catch.data,
             control = sir_control())
}
}
