% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_table}
\alias{summary_table}
\title{Function to write tables of logistic model parameter and derived quantities for StateSpaceSIR similar to Table 5 from Zerbini et al (2011).}
\usage{
summary_table(SIR, file_name = NULL)
}
\arguments{
\item{SIR}{Resample summary from StateSpaceSIR}

\item{file_name}{Desired filename to where csv file will be saved. If NULL, will not save.}
}
\description{
Function to write tables of logistic model parameter and derived quantities for StateSpaceSIR similar to Table 5 from Zerbini et al (2011).
}
