% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_dy_functions.R
\name{TARGET.K}
\alias{TARGET.K}
\title{Calculate a target K for the bisection method}
\usage{
TARGET.K(
  allee_model,
  r_max,
  K,
  N1,
  z,
  P50,
  num_Yrs,
  start_yr,
  target.Pop,
  catches,
  proc_error,
  MVP = 0
)
}
\arguments{
\item{allee_model}{Switch to determine the depensation model. 0 = no Allee effect; 1 = Hilborn et al 2014 P50 Allee Effect; 2 = Logistic Allee effect; 3 = Lin and Li 2002; 4 = Haider et al 2017.}

\item{r_max}{The maximum net recruitment rate ($r_{max}$).}

\item{K}{Pre-expoitation population size in numbers or biomass
(depending on input).}

\item{N1}{Population size in numbers or biomass at year 1 (generally
assumed to be K).}

\item{z}{Generalized logistic shape parameter, determines population
size where productivity is masimum (assumed to be 2.39 by the ISC
SC).}

\item{P50}{Parameter that determines the level of depensation.}

\item{num_Yrs}{The number of projection years. Set as the last year
in the catchor abundance series whichever is most recent, minus the
start year.}

\item{start_yr}{First year of the projection (assumed to be the first
year in the catch series).}

\item{target.Pop}{Target population size.}

\item{catches}{Catch time series. Cannot include NAs,}

\item{proc_error}{The time series of lognormal process errors. Currently
does not handle NAs or 0s}

\item{MVP}{Minimum Viable Population Size; `4 * num.haplotypes`}
}
\value{
Vector of differences between predicted population and target
  population.
}
\description{
Calculate a target K for the bisection method
}
\examples{
TARGET.K(r_max, K, N1, z, start_yr=start_yr, num_Yrs=bisection.Yrs,
         target.Pop=target.Pop, catches=catches, MVP=MVP)
}
